A reimplementation of building Quicklisp that allows circular imports between projects.
The idea mentioned in the previous commit message was implemented.
Essentially, we work in three "layers":

1. Each project has a source tarball.
2. Each system gets built to FASLs from the source tarball, but not
   tested.
3. Each project has its systems copied into a single output derivation.

The lispPackages (e.g. sbcl.packages) expose the third layer directly,
so building something like .#sbcl.packages.alexandria works as expected.

Building and testing all of Quicklisp isn't currently possible, because
the systems.txt metadata shipped by Quicklisp doesn't contain all
packages. This isn't really Quicklisp's fault; it looks like the package
clml.data.r-datasets-package is in an .asd file that can't be found,
because it's not named appropriately... A fix PR was closed in October:
https://github.com/mmaul/clml/pull/49

Overrides are also currently not implemented. They're probably the next
thing to get implemented, since they'll allow working around the
previously mentioned issue.

CFFI and UIOP also aren't building; they both die with:

    Can't create directory /nix/store/94cbpghkb4i2ijdx2fgsbg3il7sdb4ln-asdf-3.3.5/lib/uiop-3.3.5

This is from QuickLisp's copy of UIOP's bundled ASDF. Since we patch
implementations to update their ASDFs _anyway_, the easy fix might be to
special-case it to be excluded as a system dependency.