Semi-cheesy fixes.
We always use the bundled ASDF, since we're patching it to update it
further than Quicklisp's anyway; we also always prepend to
CL_SOURCE_REGISTRY and ASDF_OUTPUT_TRANSLATIONS instead of appending,
which causes some recompiling when two systems have the same .asd,
but... it's probably fine?

This gets the whole trivia project building, but coalton needs fixing
still.